refactor settings page to profile page
