--- screens/settings/SettingsScreen.js
@@ -1,20 +1,0 @@
-import * as React from "react";
-import { Button, Text, View, StyleSheet } from "react-native";
-
-export default function SettingsScreen({ signOut }) {
-  return (
-    <View style={styles.container}>
-      <Text>Signed in!</Text>
-      <Button title="Sign out" onPress={signOut} />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-});
